Модуляция Разделил код на модули. Вынес вспомогательные функции в файл functions. Завёл файлы под планируемые модули.
