+++ index.html
@@ -452,7 +452,6 @@
       </section>
     </template>
 
-    <script src="js/main.js"></script>
-    <script src="/js/function.js"></script>
+    <script src="js/main.js" type="module"></script>
   </body>
 </html>
